requirements.txt
apache-airflow-providers-apache-spark==4.8.0
pandas==2.2.2